requirements.txt
pyyaml
pytest
youtube-dl
kombu
